+++ wicket/src/cli/rack_setup/config_toml.rs
@@ -349,7 +349,7 @@ fn populate_uplink_table(cfg: &UserSpecifiedPortConfig) -> Table {
     // This style ensures that if a new field is added, this fails loudly.
     let manual_port_config = match cfg {
         UserSpecifiedPortConfig::Manual(manual) => manual,
-        UserSpecifiedPortConfig::DdmAutoPortConfig {} => {
+        UserSpecifiedPortConfig::DdmAutoPortConfig => {
             let mut uplink = Table::new();
             uplink.insert("type", string_item("ddm_auto_port_config"));
             return uplink;
